Replay Protection For Nfts (#66)
Nfts are being written over by event handlers. If the system shuts down in the middle of an update, we could be in the situation where some Nft updates have been written and others not. The event handler will need a way to know when not to apply events to a record. Events are uniquely defined and well-ordered by (block_number, log_index) which means we can prevent updates on records by introducing (last_update_block, last_update_log_index) to the record itself. This PR introduces exactly that.