Get TxData via getBlockByNumber instead of getBlockReceipts (#85)
While migrating block and TxData over to arak, I noticed that the information we needed was contained in a simpler request (when hydrated). This allows us to fetch sufficient data for blocks and transactions with a single (cheaper) request instead of 2. Currently ethrpc has a "bug/issue" that does not provide the required functionality. This has been reported to all relevant places and there is a PR in place that should resolve this