Update Test DB & Read BlockData from Arak (#90)
In alignment with the recent changes to Arak -- we update the test DB to include Blocks and Transactions (this sample file is much bigger now).

We also implement the necessary get_blocks_for_range on the read layer. This resulted in some duplicated code (namely the transactions and blocks structures in both the read layer and the data store).

This update provides arak data set for 3k blocks from block 15M to 15M + 3000.

With this, our current handler speed is ~5 Bps (Blocks per second). This will be further reduced once we offload the tokenUri, Contract (Name & Symbol) to the metadata retriever.